refactor code, rename files
